refactor(todos): migrate utility functions to a dedicated module
- Moved utility functions related to todos from the previous utils file to a new dedicated module at `_utils/todo.utils.ts`, enhancing organization and maintainability.
- Updated imports in `todo-item.tsx`, `todo-list.tsx`, and `todo-stats.tsx` to reference the new utility module, ensuring existing functionality remains intact.
- Removed the old utility file to eliminate redundancy and streamline the codebase.
